test [nfc]: Move fake-image-request code to its own file
